refactor: introduce ProofNodeClient abstraction
- Create proof_node_client.rs with HTTP transport abstraction
- Refactor proof_engine.rs to delegate to ProofNodeClient
- Update mod.rs with new module exports
- Reduces proof_engine.rs by ~185 lines
