+++ beacon_node/execution_layer/src/eip8025/proof_engine.rs
@@ -1,40 +1,28 @@
 //! ProofEngine trait and HTTP implementation for EIP-8025.
 //!
 //! This module defines the interface for interacting with proof engines
-//! and provides an HTTP REST+SSZ+SSE implementation with an internal proof cache.
+//! and provides an HTTP implementation with an internal proof cache.
 //!
-//! The proof engine backend uses a REST API with:
-//! - SSZ-encoded request bodies for proof requests
-//! - SSE (Server-Sent Events) for streaming proof completion events
-//! - HTTP endpoints for proof download and verification
+//! HTTP transport is delegated to [`super::proof_node_client::ProofNodeClient`].
 
 use super::errors::ProofEngineError;
 use super::persisted_state::PersistedProofEngineState;
+use super::proof_node_client::ProofNodeClient;
 use crate::{
     eip8025::state::{RequestMetadata, State},
     ForkchoiceState, ForkchoiceUpdatedResponse, MissingProofInfo, NewPayloadRequest,
-    NewPayloadRequestFulu, PayloadStatusV1, PayloadStatusV1Status,
+    PayloadStatusV1, PayloadStatusV1Status,
 };
 use bytes::Bytes;
-use ssz::Encode;
-use ssz_derive::Encode as SszEncode;
 use futures::stream::Stream;
 use parking_lot::RwLock;
-use reqwest::Client;
-use reqwest_eventsource::{Event, EventSource};
 use sensitive_url::SensitiveUrl;
 use std::collections::HashMap;
 use std::pin::Pin;
 use std::time::Duration;
-use tokio_stream::StreamExt;
 
 use types::execution::eip8025::{ProofAttributes, ProofStatus, SignedExecutionProof};
-use types::{EthSpec, ExecutionPayloadFulu, ExecutionRequests, Hash256, VersionedHash};
-
-use ssz_types::VariableList;
-
-/// Default timeout for proof engine requests (1 second per spec).
-pub const PROOF_ENGINE_TIMEOUT: Duration = Duration::from_secs(1);
+use types::{EthSpec, Hash256};
 
 // ─── SSE Event Types ─────────────────────────────────────────────────────────
 
@@ -116,52 +104,6 @@ impl ProofEvent {
     }
 }
 
-// ─── SSZ Helper for NewPayloadRequest ────────────────────────────────────────
-
-/// SSZ-encodable owned representation of a Fulu NewPayloadRequest.
-///
-/// Used to serialize the request body when sending to the proof engine.
-/// Field order matches the zkboost `NewPayloadRequest` Fulu variant.
-#[derive(SszEncode)]
-struct SszNewPayloadRequestFulu<E: EthSpec> {
-    execution_payload: ExecutionPayloadFulu<E>,
-    versioned_hashes: VariableList<VersionedHash, E::MaxBlobCommitmentsPerBlock>,
-    parent_beacon_block_root: Hash256,
-    execution_requests: ExecutionRequests<E>,
-}
-
-impl<'a, E: EthSpec> From<&NewPayloadRequestFulu<'a, E>> for SszNewPayloadRequestFulu<E> {
-    fn from(req: &NewPayloadRequestFulu<'a, E>) -> Self {
-        Self {
-            execution_payload: req.execution_payload.clone(),
-            versioned_hashes: req.versioned_hashes.clone(),
-            parent_beacon_block_root: req.parent_beacon_block_root,
-            execution_requests: req.execution_requests.clone(),
-        }
-    }
-}
-
-// ─── REST API Response Types ─────────────────────────────────────────────────
-
-/// Response for `POST /v1/execution_proof_requests`.
-#[derive(Debug, Clone, serde::Deserialize)]
-pub struct ProofRequestResponse {
-    pub new_payload_request_root: Hash256,
-}
-
-/// Response for `POST /v1/execution_proof_verifications`.
-#[derive(Debug, Clone, serde::Deserialize)]
-struct ProofVerificationResponse {
-    status: ProofVerificationStatus,
-}
-
-#[derive(Debug, Clone, serde::Deserialize)]
-#[serde(rename_all = "SCREAMING_SNAKE_CASE")]
-enum ProofVerificationStatus {
-    Valid,
-    Invalid,
-}
-
 // ─── ProofEngine Trait ───────────────────────────────────────────────────────
 
 /// Trait defining the interface for a proof engine.
@@ -208,18 +150,15 @@ pub trait ProofEngine: Send + Sync {
 
 // ─── HttpProofEngine ─────────────────────────────────────────────────────────
 
-/// HTTP REST+SSZ+SSE implementation of the ProofEngine trait with internal proof storage.
+/// HTTP implementation of the ProofEngine trait with internal proof storage.
 ///
 /// This implementation:
 /// - Stores ALL unfinalized proofs indexed by new_payload_request_root (unbounded)
-/// - Calls out to the proof engine REST API for proof requests and verification
-/// - Subscribes to SSE events for proof completion notifications
+/// - Delegates HTTP transport to [`ProofNodeClient`]
 /// - Prunes proofs when finalization events occur
 pub struct HttpProofEngine {
-    /// HTTP client for making requests.
-    client: Client,
-    /// URL of the proof engine endpoint.
-    url: SensitiveUrl,
+    /// Low-level HTTP client for proof engine REST+SSZ+SSE API.
+    proof_node: ProofNodeClient,
     /// The internal state storing execution proofs in a tree structure and buffer.
     state: RwLock<State>,
     /// Buffered proofs for request roots not yet seen.
@@ -227,16 +166,10 @@ pub struct HttpProofEngine {
 }
 
 impl HttpProofEngine {
-    /// Create a new HTTP proof engine client with internal proof storage.
+    /// Create a new HTTP proof engine with internal proof storage.
     pub fn new(url: SensitiveUrl, timeout: Option<Duration>) -> Self {
-        let client = Client::builder()
-            .timeout(timeout.unwrap_or(PROOF_ENGINE_TIMEOUT))
-            .build()
-            .expect("Failed to build HTTP client");
-
         Self {
-            client,
-            url,
+            proof_node: ProofNodeClient::new(url, timeout),
             state: RwLock::new(State::new()),
             buffered_proofs: RwLock::new(HashMap::new()),
         }
@@ -244,59 +177,25 @@ impl HttpProofEngine {
 
     /// Subscribe to SSE proof events from the proof engine.
     ///
-    /// Opens `GET /v1/execution_proof_requests` as an SSE stream.
-    /// When `filter_root` is provided, only events for that root are received.
+    /// Delegates to [`ProofNodeClient::subscribe_proof_events`].
     pub fn subscribe_proof_events(
         &self,
         filter_root: Option<Hash256>,
     ) -> Pin<Box<dyn Stream<Item = Result<ProofEvent, ProofEngineError>> + Send + '_>> {
-        let client = self.client.clone();
-        let base_url = self.url.expose_full().clone();
-
-        Box::pin(async_stream::try_stream! {
-            let mut url = base_url;
-            url.set_path("/v1/execution_proof_requests");
-            if let Some(root) = filter_root {
-                url.set_query(Some(&format!("new_payload_request_root={root}")));
-            }
-
-            let builder = client.get(url);
-            let mut es = EventSource::new(builder)
-                .map_err(|e| ProofEngineError::SseError(
-                    format!("failed to create event source: {e}")
-                ))?;
-
-            while let Some(event) = es.next().await {
-                match event {
-                    Ok(Event::Open) => {}
-                    Ok(Event::Message(message)) => {
-                        yield ProofEvent::try_from_parts(&message.event, &message.data)?;
-                    }
-                    Err(error) => {
-                        es.close();
-                        Err(ProofEngineError::SseError(error.to_string()))?;
-                    }
-                }
-            }
-        })
+        self.proof_node.subscribe_proof_events(filter_root)
     }
 
     /// Download a completed execution proof by proof type.
     ///
-    /// Sends `GET /v1/execution_proofs/{root}/{proof_type}`.
+    /// Delegates to [`ProofNodeClient::get_proof`].
     pub async fn get_proof(
         &self,
         new_payload_request_root: Hash256,
         proof_type: u8,
     ) -> Result<Bytes, ProofEngineError> {
-        let mut url = self.url.expose_full().clone();
-        url.set_path(&format!(
-            "/v1/execution_proofs/{new_payload_request_root}/{proof_type}"
-        ));
-
-        let response = self.client.get(url).send().await?.error_for_status()?;
-
-        Ok(response.bytes().await?)
+        self.proof_node
+            .get_proof(new_payload_request_root, proof_type)
+            .await
     }
 
     /// Snapshot the current state into a persisted form for serialization.
@@ -309,75 +208,6 @@ impl HttpProofEngine {
     pub fn restore_from_persisted(&self, persisted: PersistedProofEngineState) {
         let restored = persisted.to_state();
         *self.state.write() = restored;
-    }
-
-    /// Send a proof request to the proof engine REST API.
-    ///
-    /// `POST /v1/execution_proof_requests?proof_types=0,1,2`
-    /// Body: SSZ-encoded NewPayloadRequest
-    async fn request_proofs_rest<E: EthSpec>(
-        &self,
-        new_payload_request_fulu: NewPayloadRequestFulu<'_, E>,
-        proof_attributes: ProofAttributes,
-    ) -> Result<Hash256, ProofEngineError> {
-        let mut url = self.url.expose_full().clone();
-        url.set_path("/v1/execution_proof_requests");
-
-        let proof_types_str = proof_attributes
-            .proof_types
-            .iter()
-            .map(|t| t.to_string())
-            .collect::<Vec<_>>()
-            .join(",");
-        url.set_query(Some(&format!("proof_types={proof_types_str}")));
-
-        let ssz_body = SszNewPayloadRequestFulu::from(&new_payload_request_fulu);
-
-        let response: ProofRequestResponse = self
-            .client
-            .post(url)
-            .header("content-type", "application/octet-stream")
-            .body(ssz_body.as_ssz_bytes())
-            .send()
-            .await?
-            .error_for_status()?
-            .json()
-            .await?;
-
-        Ok(response.new_payload_request_root)
-    }
-
-    /// Verify a proof via the proof engine REST API.
-    ///
-    /// `POST /v1/execution_proof_verifications?new_payload_request_root=...&proof_type=...`
-    /// Body: raw proof bytes
-    async fn verify_proof_rest(
-        &self,
-        new_payload_request_root: Hash256,
-        proof_type: u8,
-        proof_data: &[u8],
-    ) -> Result<ProofStatus, ProofEngineError> {
-        let mut url = self.url.expose_full().clone();
-        url.set_path("/v1/execution_proof_verifications");
-        url.set_query(Some(&format!(
-            "new_payload_request_root={new_payload_request_root}&proof_type={proof_type}"
-        )));
-
-        let response: ProofVerificationResponse = self
-            .client
-            .post(url)
-            .header("content-type", "application/octet-stream")
-            .body(proof_data.to_vec())
-            .send()
-            .await?
-            .error_for_status()?
-            .json()
-            .await?;
-
-        match response.status {
-            ProofVerificationStatus::Valid => Ok(ProofStatus::Valid),
-            ProofVerificationStatus::Invalid => Ok(ProofStatus::Invalid),
-        }
     }
 }
 
@@ -414,7 +244,8 @@ impl ProofEngine for HttpProofEngine {
         }
 
         let status = self
-            .verify_proof_rest(
+            .proof_node
+            .verify_proof(
                 proof.request_root(),
                 proof.proof_type(),
                 &proof.message.proof_data,
@@ -482,7 +313,8 @@ impl ProofEngine for HttpProofEngine {
                 Err(ProofEngineError::ForkNotSupported("Electra".to_string()))
             }
             NewPayloadRequest::Fulu(new_payload_request_fulu) => {
-                self.request_proofs_rest(new_payload_request_fulu, proof_attributes)
+                self.proof_node
+                    .request_proofs(new_payload_request_fulu, proof_attributes)
                     .await
             }
             NewPayloadRequest::Gloas(_) => {
